+++ src/chains/index.ts
@@ -1,5 +1,5 @@
 export * from './Chain'
+export * from './chain.utils'
 export * from './EVMChain'
 export * from './SolanaChain'
 export * from './supported.chains'
-export * from './chain.utils'