+++ src/chains/supported.chains.ts
@@ -131,8 +131,8 @@ export const supportedEVMChains: EVMChain[] = [
       },
       rpcUrls: [
         'https://rpc.gnosischain.com/',
-        'https://rpc.xdaichain.com/',
-        'https://dai.poa.network/',
+        'https://rpc.ankr.com/gnosis',
+        'https://xdai-rpc.gateway.pokt.network',
       ],
     },
   },
@@ -1332,6 +1332,31 @@ export const supportedEVMChains: EVMChain[] = [
         decimals: 18,
       },
       rpcUrls: ['https://api.avax-test.network/ext/bc/C/rpc'],
+    },
+  },
+  // 59140 - ConsenSys zkEVM Testnet
+  {
+    key: ChainKey.CZKT,
+    chainType: ChainType.EVM,
+    name: 'ConsenSys zkEVM Testnet',
+    coin: CoinKey.ETH,
+    id: 59140,
+    mainnet: false,
+    logoURI:
+      'https://raw.githubusercontent.com/lifinance/types/main/src/assets/icons/chains/consensys_zkevm_test.png',
+    // multicallAddress: multicallAddresses[ChainId.CZKT], // Not deployed yet
+    faucetUrls: ['https://goerli.zkevm.consensys.net/'],
+
+    metamask: {
+      chainId: prefixChainId(59140),
+      blockExplorerUrls: ['https://explorer.goerli.zkevm.consensys.net'],
+      chainName: 'ConsenSys zkEVM Testnet',
+      nativeCurrency: {
+        name: 'crETH',
+        symbol: 'crETH',
+        decimals: 18,
+      },
+      rpcUrls: [],
     },
   },
   // TODO: Add