+++ src/coins.ts
@@ -47,7 +47,7 @@ const basicCoins: BasicCoin[] = [
         address: '0x6a023ccd1ff6f2045c3309768ead9e68f978f6e1',
         decimals: 18,
         symbol: 'WETH',
-        name: 'Wrapped Ether on xDai',
+        name: 'Wrapped Ether',
       },
       [ChainId.OPT]: {
         address: '0x0000000000000000000000000000000000000000',
@@ -142,6 +142,14 @@ const basicCoins: BasicCoin[] = [
         address: '0x0000000000000000000000000000000000000000',
         decimals: 18,
       },
+      [ChainId.ARBG]: {
+        address: '0x0000000000000000000000000000000000000000',
+        decimals: 18,
+      },
+      [ChainId.CZKT]: {
+        address: '0x0000000000000000000000000000000000000000',
+        decimals: 18,
+      },
     },
   },
   // > MATIC
@@ -173,7 +181,7 @@ const basicCoins: BasicCoin[] = [
       [ChainId.DAI]: {
         address: '0x7122d7661c4564b7c6cd4878b06766489a6028a2',
         decimals: 18,
-        name: 'Matic Token on xDai',
+        name: 'Matic Token',
       },
       // https://evmexplorer.velas.com/token/0x6ab0B8C1a35F9F4Ce107cCBd05049CB1Dbd99Ec5/
       [ChainId.VEL]: {
@@ -217,7 +225,7 @@ const basicCoins: BasicCoin[] = [
         address: '0xca8d20f3e0144a72c6b5d576e9bd3fd8557e2b04',
         decimals: 18,
         symbol: 'WBNB',
-        name: 'Wrapped BNB on xDai',
+        name: 'Wrapped BNB',
       },
       [ChainId.ONE]: {
         address: '0xb1f6e61e1e113625593a22fa6aa94f8052bc39e0',
@@ -248,7 +256,7 @@ const basicCoins: BasicCoin[] = [
   // > DAI
   {
     key: CoinKey.DAI,
-    name: 'Dai Stablecoin',
+    name: 'DAI Stablecoin',
     logoURI:
       'https://raw.githubusercontent.com/trustwallet/assets/master/blockchains/ethereum/assets/0x6B175474E89094C44Da98b954EedeAC495271d0F/logo.png',
     verified: true,
@@ -260,23 +268,22 @@ const basicCoins: BasicCoin[] = [
       [ChainId.SOL]: {
         address: 'EjmyN6qEC1Tf1JxiG1ae7UTJhUxSwk1TCWNWqxWV4J6o',
         decimals: 8,
-        name: 'Dai Stablecoin (Wormhole)',
+        name: 'DAI Stablecoin (Wormhole)',
       },
       [ChainId.BSC]: {
         address: '0x1af3f329e8be154074d8769d1ffa4ee058b1dbc3',
         decimals: 18,
-        name: 'Dai Token',
       },
       [ChainId.POL]: {
         address: '0x8f3cf7ad23cd3cadbd9735aff958023239c6a063',
         decimals: 18,
-        name: '(PoS) Dai Stablecoin',
+        name: '(PoS) DAI Stablecoin',
       },
       [ChainId.DAI]: {
         address: '0x0000000000000000000000000000000000000000',
         decimals: 18,
         symbol: 'xDAI',
-        name: 'xDai Native Token',
+        name: 'xDAI Native Token',
       },
       [ChainId.OPT]: {
         address: '0xda10009cbd5d07dd0cecc66161fc93d7c9000da1',
@@ -312,7 +319,7 @@ const basicCoins: BasicCoin[] = [
       [ChainId.FUS]: {
         address: '0x94ba7a27c7a95863d1bdc7645ac2951e0cca06ba',
         decimals: 18,
-        name: 'Dai Stablecoin on Fuse',
+        name: 'DAI Stablecoin',
       },
       [ChainId.CEL]: {
         address: '0x90ca507a5d4458a4c6c6249d186b6dcb02a5bccd',
@@ -408,6 +415,10 @@ const basicCoins: BasicCoin[] = [
         address: 'KgV1GvrHQmRBY8sHQQeUKwTm2r2h8t4C8qt12Cw1HVE',
         decimals: 8,
         name: 'Avalanche (Wormhole)',
+      },
+      [ChainId.AVAT]: {
+        address: '0x0000000000000000000000000000000000000000',
+        decimals: 18,
       },
     },
   },
@@ -713,6 +724,11 @@ const basicCoins: BasicCoin[] = [
         address: '0x6d4dd09982853f08d9966ac3ca4eb5885f16f2b2',
         decimals: 6,
       },
+      [ChainId.CZKT]: {
+        address: '0x964FF70695da981027c81020B1c58d833D49A640',
+        decimals: 6,
+        name: 'ConsenSys USD Coin',
+      },
       // 42, 0xb7a4f3e9097c08da09517b5ab877f7a917224ede, 6
     },
   },
@@ -864,7 +880,7 @@ const basicCoins: BasicCoin[] = [
         address: '0x6a023ccd1ff6f2045c3309768ead9e68f978f6e1',
         decimals: 18,
         symbol: 'WETH',
-        name: 'Wrapped Ether on xDai',
+        name: 'Wrapped Ether',
       },
       [ChainId.POL]: {
         address: '0x7ceb23fd6bc0add59e62ac25578270cff1b9f619',
@@ -954,6 +970,10 @@ const basicCoins: BasicCoin[] = [
       },
       [ChainId.OPTG]: {
         address: '0x4200000000000000000000000000000000000006',
+        decimals: 18,
+      },
+      [ChainId.CZKT]: {
+        address: '0x2C1b868d6596a18e32E61B901E4060C872647b6C',
         decimals: 18,
       },
     },
@@ -1586,6 +1606,17 @@ export const wrappedTokens: { [ChainId: string]: Token } = {
     logoURI:
       'https://zapper.fi/images/networks/ethereum/0x0000000000000000000000000000000000000000.png',
   },
+  [ChainId.ARBG]: {
+    // https://goerli.arbiscan.io/token/0x42da9eE191833756c618778145A86E6709f70C9b
+    address: '0x42da9eE191833756c618778145A86E6709f70C9b',
+    symbol: 'WETH',
+    decimals: 18,
+    chainId: ChainId.ARBG,
+    coinKey: CoinKey.WETH,
+    name: 'WETH',
+    logoURI:
+      'https://zapper.fi/images/networks/ethereum/0x0000000000000000000000000000000000000000.png',
+  },
   [ChainId.OPTT]: {
     // https://kovan-optimistic.etherscan.io/token/0x4200000000000000000000000000000000000006
     address: '0x4200000000000000000000000000000000000006',
@@ -1618,6 +1649,27 @@ export const wrappedTokens: { [ChainId: string]: Token } = {
     name: 'WBNB',
     logoURI:
       'https://zapper.fi/images/networks/binance-smart-chain/0x0000000000000000000000000000000000000000.png',
+  },
+  [ChainId.CZKT]: {
+    // https://explorer.prealpha.zkevm.consensys.net/address/0x2C1b868d6596a18e32E61B901E4060C872647b6C
+    address: '0x2C1b868d6596a18e32E61B901E4060C872647b6C',
+    symbol: 'WETH',
+    decimals: 18,
+    chainId: ChainId.CZKT,
+    coinKey: CoinKey.WETH,
+    name: 'WETH',
+    logoURI:
+      'https://zapper.fi/images/networks/ethereum/0x0000000000000000000000000000000000000000.png',
+  },
+  [ChainId.AVAT]: {
+    // https://testnet.snowtrace.io/token/0xd00ae08403B9bbb9124bB305C09058E32C39A48c
+    address: '0xd00ae08403B9bbb9124bB305C09058E32C39A48c',
+    symbol: 'WAVAX',
+    decimals: 18,
+    chainId: ChainId.AVA,
+    coinKey: 'WAVAX' as CoinKey,
+    name: 'Wrapped AVAX',
+    logoURI: '',
   },
 }